write prompt for AI, fetch AI data on form submit, display data on details page, values of form fields will be used in next commit
